backend/pyproject.toml
[tool.poetry]
name = "model-lab-backend"
version = "0.1.0"
description = "Backend for Model Lab - Multi-model LLM API"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.103.1"
uvicorn = "^0.23.2"
openai = "^1.3.0"
python-dotenv = "^1.0.0"
pydantic = "^2.4.2"

[tool.poetry.dev-dependencies]
pytest = "^7.0.0"
black = "^23.3.0"
isort = "^5.12.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"



